Preliminary game logic and small refactoring
