+++ Ohshie.FlashCards/CardsManager/FlashcardEditor.cs
@@ -5,22 +5,7 @@ namespace Ohshie.FlashCards.CardsManager;
 
 public class FlashcardEditor
 {
-    private readonly DbOperations _dbOperations = new DbOperations();
-    
-    // public (string, string) RewriteFlashcard(FlashCardDto flashCardDto)
-    // {
-    //     var updatedFlashcardName = AskUser.AskTextInput(where: "flashcard", what: "name");
-    //     var updatedFlashcardContent = AskUser.AskTextInput(where: "flashcard", what: "content");
-    //
-    //     var flashcard = Mapper.FlashcardDtoToFlashcardMapper(flashCardDto);
-    //     
-    //     _dbOperations.UpdateFlashcard(flashcard, updatedFlashcardName, updatedFlashcardContent);
-    //
-    //     flashCardDto.Name = updatedFlashcardName;
-    //     flashCardDto.Content = updatedFlashcardContent;
-    //     
-    //     return (updatedFlashcardName, updatedFlashcardContent);
-    // }
+    private readonly FlashcardsRepository _flashcardsRepository = new();
     
     public void RewriteFlashcard(FlashCardDto flashCardDto)
     {
@@ -29,7 +14,7 @@ public class FlashcardEditor
 
         var flashcard = Mapper.FlashcardDtoToFlashcardMapper(flashCardDto);
         
-        _dbOperations.UpdateFlashcard(flashcard, updatedFlashcardName, updatedFlashcardContent);
+        _flashcardsRepository.UpdateFlashcard(flashcard, updatedFlashcardName, updatedFlashcardContent);
 
         flashCardDto.Name = updatedFlashcardName;
         flashCardDto.Content = updatedFlashcardContent;
