+++ Ohshie.FlashCards/DataAccess/SessionRepository.cs
@@ -1,0 +1,6 @@
+namespace Ohshie.FlashCards.DataAccess;
+
+public class SessionRepository : DbBase
+{
+    
+}
