--- Ohshie.FlashCards/Menus/GameMenu.cs
@@ -1,25 +1,0 @@
-namespace Ohshie.FlashCards.Menus;
-
-public class GameMenu : MenuBase
-{
-    protected override string[] MenuItems { get; } =
-    {
-        "Placeholder"
-    };
-
-    public new void Initialize()
-    {
-        if(!Verify.DeckExist()) return;
-        
-        bool chosenExit = false;
-        while (!chosenExit)
-        {
-            chosenExit = Menu();
-        }
-    }
-    
-    protected override bool Menu()
-    {
-        throw new NotImplementedException();
-    }
-}
