+++ Ohshie.FlashCards/Verify.cs
@@ -1,16 +1,18 @@
+using Ohshie.FlashCards.DataAccess;
+
 namespace Ohshie.FlashCards;
 
 public static class Verify
 {
-    public static bool DeckExist()
+    public static bool DeckExist(string where)
     {
-        DbOperations dbOperations = new();
-        var stacks = dbOperations.FetchAllDecks();
+        DecksRepository decksRepository = new();
+        var stacks = decksRepository.FetchAllDecks();
         
         if (!stacks.Any())
         {
             AnsiConsole.MarkupLine("Oops, you got no [bold]Decks[/]. Please create one first\n" +
-                                  "Press enter to go back");
+                                  $"Press enter to {where}");
             Console.ReadLine();
 
             return false;
